Made require('store2') possible from node.js.
Requiring this module resulted in an 'ReferenceError: window is not defined' error.
Using `this` as reference to the global scope / `window` instead fixes this issue.

You cant actually use this module in node, as there is no such thing as `localStorage` or `sessionStorage`.
Although this enables isomorphic webapps to require('store2') in parts of the codebase which is *shared* by server and client.